sepolicy: Whitelist system applications in netd context
Some system applications require a permission to find data through netd
service.  As all requsts for find operation from service manager are
getting thrown after audition, it is obligatory to set an allow for that
purpose, but this was not possible due to a neverallow filter set in a
genetic SE policy.

Allow system applications a find operation to avoid security fails on
a number of affected builds.

Change-Id: Ifcfcae0d5be127e0a15459f3acb5badd1c3673fe
Signed-off-by: Alex Saiko <solcmdr@gmail.com>